requirements.txt
setuptools
groq
flask
python-dotenv
flask-cors
flask-socketio
pymongo
werkzeug
pydub
speechrecognition
sounddevice
numpy
edge-tts
numpy
webrtcvad
asyncio
pygame
langdetect